requirements.txt
fastapi==0.111.0
keras==2.15.0
numpy==1.26.4
pillow==10.3.0
uvicorn==0.30.1
tensorflow==2.15.0